workshop-2024-06-23/telegram-bot/requirements.txt
python-telegram-bot==21.3
openai==1.35.3